feat: centralize data and output directories, activate environment variables, and update base repository to master branch.

+++ docker-entrypoint.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 set -e
+[ ! -L "data/GPX_OUT" ] && rm -rf GPX_OUT && mkdir "data/GPX_OUT"; ln -rs data/GPX_OUT GPX_OUT
+[ ! -L "data/FIT_OUT" ] && rm -rf FIT_OUT && mkdir "data/FIT_OUT"; ln -rs data/FIT_OUT FIT_OUT
+[ ! -L "data/TCX_OUT" ] && rm -rf TCX_OUT && mkdir "data/TCX_OUT"; ln -rs data/TCX_OUT TCX_OUT
+[ ! -L "data/dist" ] && rm -rf dist && mkdir "data/dist"; ln -rs data/dist dist
 
 if [[ -n $APP && ! -f "data/.live" ]]; then
     echo "Running initialization tasks for $APP..."
@@ -11,7 +15,7 @@ if [[ -n $APP && ! -f "data/.live" ]]; then
     ln -rs data/activities.json src/static/activities.json
     ln -rs data/imported.json imported.json
     ln -rs data/data.db run_page/data.db
-elif [[ -z $APP && ! -f "data/.demo" ]]; then
+elif [[ -z $APP && ! -f "data/.demo" && ! -f "data/.live" ]]; then
     echo "Running demo initializaion..."
     curl -o src/static/activities.json https://raw.githubusercontent.com/yihong0618/running_page/refs/heads/master/src/static/activities.json
     echo -n | tee data/{imported.json,data.db,.demo}
